+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## []
 
 - Added Trade info support for vanilla
+- Fixed memory leak in JEI plugin
 
 ## [1.2.5]